Restore diff support
