+++ identity-core/src/common/url.rs
@@ -9,6 +9,9 @@ use core::ops::Deref;
 use core::ops::DerefMut;
 use core::str::FromStr;
 
+use crate::diff;
+use crate::diff::Diff;
+use crate::diff::DiffString;
 use crate::error::Error;
 use crate::error::Result;
 
@@ -61,6 +64,12 @@ impl DerefMut for Url {
   }
 }
 
+impl From<::url::Url> for Url {
+  fn from(other: ::url::Url) -> Self {
+    Self(other)
+  }
+}
+
 impl FromStr for Url {
   type Err = Error;
 
@@ -75,5 +84,28 @@ where
 {
   fn eq(&self, other: &T) -> bool {
     self.as_str() == other.as_ref()
+  }
+}
+
+impl Diff for Url {
+  type Type = DiffString;
+
+  fn diff(&self, other: &Self) -> diff::Result<Self::Type> {
+    self.to_string().diff(&other.to_string())
+  }
+
+  fn merge(&self, diff: Self::Type) -> diff::Result<Self> {
+    self
+      .to_string()
+      .merge(diff)
+      .and_then(|this| Self::parse(&this).map_err(diff::Error::merge))
+  }
+
+  fn from_diff(diff: Self::Type) -> diff::Result<Self> {
+    String::from_diff(diff).and_then(|this| Self::parse(&this).map_err(diff::Error::convert))
+  }
+
+  fn into_diff(self) -> diff::Result<Self::Type> {
+    self.to_string().into_diff()
   }
 }
