+++ identity-did/Cargo.toml
@@ -12,7 +12,7 @@ homepage = "https://www.iota.org"
 
 [dependencies]
 async-trait = { version = "0.1", default-features = false }
-did_url = { version = "0.1", default-features = false, features = ["alloc", "serde"] }
+did_url = { version = "0.1", default-features = false, features = ["std", "serde"] }
 identity-core = { version = "=0.1.0", path = "../identity-core" }
 serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
 thiserror = { version = "1.0", default-features = false }
