+++ identity-did/src/diff/mod.rs
@@ -1,0 +1,17 @@
+// Copyright 2020-2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod did_key;
+mod document;
+mod method;
+mod method_data;
+mod method_ref;
+mod method_type;
+mod ordered_set;
+mod service;
+
+pub use self::document::DiffDocument;
+pub use self::method::DiffMethod;
+pub use self::method_data::DiffMethodData;
+pub use self::method_ref::DiffMethodRef;
+pub use self::service::DiffService;
